requirements.txt

certifi==2020.6.20
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.20
tweepy==4.5.0
Werkzeug==1.0.1
wincertstore==0.2
openpyxl==3.0.5
facebook-sdk==3.1.0
python-dotenv
yelpapi
pandas
pynytimes



